refactor(rest): move power validation to model util
