+++ rest-api/api/pkg/api/handler/instance.go
@@ -547,7 +547,7 @@ func (cih CreateInstanceHandler) Handle(c echo.Context) error {
 		logger.Warn().Msg(fmt.Sprintf("The Site: %v where this Instance is being created is not in Registered state", vpc.SiteID.String()))
 		return cutil.NewAPIErrorResponse(c, http.StatusBadRequest, "The Site where this Instance is being created is not in Registered state", nil)
 	}
-	if apiErr := common.ValidateSitePowerManagement(site.Config, apiRequest.PowerProfile); apiErr != nil {
+	if apiErr := util.ValidateSitePowerManagement(site.Config, apiRequest.PowerProfile); apiErr != nil {
 		return cutil.NewAPIErrorResponse(c, apiErr.Code, apiErr.Message, nil)
 	}
 	// Begin validating interfaces
@@ -2487,7 +2487,7 @@ func (uih UpdateInstanceHandler) Handle(c echo.Context) error {
 		logger.Error().Str("Site ID", site.ID.String()).Str("Site Status", site.Status).Msg("Unable to update Instance, Site is not in Registered state")
 		return cutil.NewAPIErrorResponse(c, http.StatusBadRequest, "Site is not in Registered state - cannot update Instance", nil)
 	}
-	if apiErr := common.ValidateSitePowerManagement(site.Config, apiRequest.PowerProfile); apiErr != nil {
+	if apiErr := util.ValidateSitePowerManagement(site.Config, apiRequest.PowerProfile); apiErr != nil {
 		return cutil.NewAPIErrorResponse(c, apiErr.Code, apiErr.Message, nil)
 	}
 	// If the instance is in some stage of deprovisioning, there's nothing to update.
